backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
groq==0.4.1
pandas==2.1.4
mysql-connector-python==8.2.0
reportlab==4.0.9
matplotlib==3.8.2
Pillow==10.2.0
python-multipart==0.0.6
pydantic==2.5.3
